Upgrade of FHIR .NET API as follows: Hl7.Fhir.STU3.Core, Version=0.93.4.0  to  Hl7.Fhir.STU3.Core, Version=0.94.0.0 Hl7.Fhir.Support, Version=0.4.2.0    to   Hl7.Fhir.Support, Version=0.94.0.0 Hl7.FhirPath, Version=0.4.2.0   to   Hl7.FhirPath, Version=0.94.0.0
Number of fixes to bring in new version:
Refactor how datetime works:
There was a change in the FHIR API in regard to how PartialDateTime.ToUniversalTime() worked. This caused an issue and made we refactor date and Time usage significantly. I now only store Zulu time zone '00:00' in the database rather than the local time zone. This makes the database no longer relevant to the physical location of the server. Yet it is not as simple as saying it now is no relevant. When the server is given a Datetime that has no time zone the server has to assume a time zone. It will assume the local system time zone of the server and then converts the datetime based on local time zone to zulu time for storage in the database. This means that the database is still dependant of it's location for any datetime that had no timezone, as it has assumed server local time. It the server was moved to a different timezone then these database entries would effectively also slide to the new timezone. They would move in time. Anything that had a timezone would be unaffected. This is also true for an datetime search query with no timezone. Once again the server must assume a time zone and it uses the local timezone in this case. The wrap of this change means that this is a braking change as it will render all databases entries before the change incorrect post the chnage as the times will all move. The database must be dropped and start from scratch. Unless a script was written to  convert all dateTimes in the database from their current timezone to zulu time. At least moving forward the resoning is simpler as all inbound datetimes are converted to zulu time and the server works only in Zulu time. The actual FHIR resource content is not effected or changed only the database indexes.
XML and Json serialisation
XML and Json serialisation changed in FHIR API, uses new object to do this.

ETags
I assume something changed as all my test cases here broke, yet I also feel I had a bug and the change just exposed it. All working correctly now.

FHIRPath
Small change to how FHIRPath Navigator is used.
PocoNavigator changed to namespace from hl7.fhir.FhirPath to hl7.fhir.ElementModel

Static string variable for Validate moved from hl7.fhir.FhirClientOperations to hl7.fhir.rest.restOperation.VALIDATE_RESOURCE

This now all compiles and all test cases pass. Yet I now need to check the code generation works correctly and see if anything new is generated. That will be the next commit.