Fixed bug found in chain searches where the chained parameter uses a search parameter that is common to all resource types, such as the '_id' search parameter. For instance a search chains like these: Observation?subject:Patient._id=FCC-PAT-0001 or even this: DiagnosticReport?result:Observation.subject:Patient._id=FCC-PAT-00001
This now works where it was an error before stating that 'Resource' is not a supported Resource Type. This was  because in the Pyro server the search parameters for all resources types have a resource type of 'Resource' set on them by default. We needed to get the Modifier Type resource and use this, i.e Patient and Observation in the examples above.