This commit implements support for search parameters of type Composite, the last type to be supported. To support this a change had to be made to the first time start-up of the server. Now, when you perform a clean install and and then start the API for the first time, a background FHIR task in launched which loads all the specifications SearchParamter resources, setting each as active indexes in the server. During this process the server is set into a read-only mode not accepting any POST or PUT actions. This is to prevent resources from being committed before the search parameters for that resource are loaded, which would result in that resource's search parameters not being indexed. The Task will take roughly 10min to run and will never need to be run again. You can monitor the progress by attempting to PUT or POST a resource and an OperationOutcome will be returned reporting the progress on the Task. Once the Task is finished the server will switch out of read-only mode. This process will also run if you are applying this upgrade to an existing server.
There are now two new PyroApp.config commands to allow the administrator to put the server in a Read-Only mode and to return a custom message explaining why. The commands are 'ServerReadOnlyMode' True/False and 'ServerReadOnlyModeMessage' which is a string. Please read to full documnetation found in the Pyro.WebAPI PyroApp.config file.

This commit also changes the way indexes are deleted for history resource. Before this commit the indexes were not deleted as a new resource instances were committed, rather an operation named  '$server-indexes-delete-history-indexes' had to be called to perform the delete action. This has now changed and the indexes are now deleted as the new resource is committed, there is no need to call the operation and furthermore the operation has been removed from the server.

This commit has database changes and the Pyro. DbManager must be run.
RevVersion3.0.2.12